Eliminate "useless cast" warnings related to counting functions.
Running GCC with the "useless-cast" warning enabled produces a number
of warnings.  Functions like count(), size(), length(), width() and
height() typically already return and int and don't need to be
explicitly cast to one.

+++ mythplugins/mytharchive/mytharchive/mythburn.cpp
@@ -290,7 +290,7 @@ void MythBurn::loadEncoderProfiles()
     QDomNodeList profileNodeList = doc.elementsByTagName("profile");
     QString name, desc, bitrate;
 
-    for (int x = 0; x < (int) profileNodeList.count(); x++)
+    for (int x = 0; x < profileNodeList.count(); x++)
     {
         QDomNode n = profileNodeList.item(x);
         QDomElement e = n.toElement();
