+++ mythtv/libs/libmythtv/eitfixup.cpp
@@ -750,7 +750,7 @@ void EITFixUp::SetUKSubtitle(DBEventEIT &event) const
              uint nTitle=0;
              int nTitleMax=-1;
              int i;
-             for (i =0; (i<(int)strListColon.count()) && (nTitleMax==-1);i++)
+             for (i =0; (i<strListColon.count()) && (nTitleMax==-1);i++)
              {
                  const QStringList tmp = strListColon[i].split(" ");
 
@@ -769,7 +769,7 @@ void EITFixUp::SetUKSubtitle(DBEventEIT &event) const
                  strPartial+=strListTmp[nTitleMax-1];
                  strListEnd.push_back(strPartial);
              }
-             for (i=nTitleMax+1;i<(int)strListColon.count();i++)
+             for (i=nTitleMax+1;i<strListColon.count();i++)
                  strListEnd.push_back(strListColon[i]);
              fColon = true;
          }
@@ -1122,7 +1122,7 @@ void EITFixUp::FixUK(DBEventEIT &event) const
 
     // Trim leading/trailing '.'
     event.subtitle.remove(m_ukDotSpaceStart);
-    if (event.subtitle.lastIndexOf("..") != (((int)event.subtitle.length())-2))
+    if (event.subtitle.lastIndexOf("..") != (event.subtitle.length()-2))
         event.subtitle.remove(m_ukDotEnd);
 
     // Reverse the subtitle and empty description
@@ -1612,7 +1612,7 @@ void EITFixUp::FixMCA(DBEventEIT &event) const
 
     // Dolby Digital 5.1?
     position = event.description.indexOf(m_mcaDD);
-    if ((position > 0) && (position > (int) (event.description.length() - 7)))
+    if ((position > 0) && (position > event.description.length() - 7))
     {
         event.audioProps |= AUD_DOLBY;
         event.description.replace(m_mcaDD, "");
